# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: WOVSDH Nails Booking API
  description: |
    Complete RESTful API for online nail salon booking system with Telegram notifications, Google Calendar integration, and email confirmations.

    ## Base URLs

    - **Local Development**: `http://localhost:3000`
    - **Production**: `https://your-domain.com`

    ## Authentication

    ### Public Endpoints
    Most endpoints are public and don't require authentication. However, Supabase RLS (Row Level Security) ensures data isolation.

    ### Admin Endpoints
    Admin endpoints require `X-Admin-Secret` header with the value of `ADMIN_SECRET_KEY` environment variable.

    ### Internal Endpoints
    Internal endpoints (confirmed bookings, email sending) require `X-Internal-Secret` header with the value of `ADMIN_SECRET_KEY`.

    ## Response Format

    All responses are JSON. Errors follow a standard format:

    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message"
      }
    }
    ```

    ## Dates and Times

    All timestamps are returned in ISO 8601 format with timezone offset (e.g., `2025-06-01T08:00:00+03:00`).

    All dates in query parameters use format `YYYY-MM-DD`.

    ## Rate Limiting

    No rate limiting is currently implemented. For production use, implement rate limiting via Vercel's edge functions or external service.

    ## Webhooks

...
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
  /api/slots:
    get:
      summary: Get available and booked slots for a specific date
      description: |
        Retrieve all slots (available, booked, blocked) for a given studio and date.

        Returns both free slots and booked slots with associated booking information.

        Results are paginated by date.
      operationId: getSlots
      tags:
        - Slots
      parameters:
        - name: studio_id
          in: query
          required: true
          schema:
            type: string
            enum:
              - rishon
              - ashdod
          description: Studio identifier
          example: rishon
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD format (local Asia/Jerusalem time)
          example: '2025-06-01'
      responses:
        '200':
          description: Successfully retrieved slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSlotsResponse'
          examples:
            success:
              value:
                slots:
                  - id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                    studio_id: rishon
                    start_at: '2025-06-01T08:00:00+03:00'
                    end_at: '2025-06-01T09:00:00+03:00'
                    status: available
                  - id: 7f8c72c0-1234-5678-9abc-def012345678
                    studio_id: rishon
                    start_at: '2025-06-01T09:00:00+03:00'
                    end_at: '2025-06-01T10:00:00+03:00'
                    status: booked
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                invalid_params:
                  value:
                    error:
                      code: INVALID_PARAMS
                      message: Invalid studio_id or date format
        '404':
          description: Studio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                not_found:
                  value:
                    error:
                      code: SLOT_NOT_FOUND
                      message: No studio found with id 'invalid'
...
  /api/internal/confirm-booking:
    post:
      summary: Confirm a pending booking (internal)
      description: |
        Internal endpoint to confirm a booking. This is called by Telegram webhook when staff clicks "Confirm" button.

        Confirmation will:
        1. Update booking status to CONFIRMED
        2. Create event in Google Calendar
        3. Send confirmation email to client
        4. Update Telegram message

        Requires internal secret authentication.
      operationId: confirmBooking
      tags:
        - Internal
      security:
        - InternalSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                booking_id:
                  type: string
                  format: uuid
                  description: Booking ID to confirm
              required:
                - booking_id
            examples:
              example1:
                value:
                  booking_id: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Booking confirmed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  event_id:
                    type: string
                    description: Google Calendar event ID
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/internal/send-confirmation-email:
    post:
      summary: Send confirmation email (internal)
      description: |
        Internal endpoint to send confirmation email to a client. Normally called automatically by the confirmation flow, but can be manually triggered.

        Requires internal secret authentication.
      operationId: sendConfirmationEmail
      tags:
        - Internal
      security:
        - InternalSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                booking_id:
                  type: string
                  format: uuid
                  description: Booking ID
              required:
                - booking_id
            examples:
              example1:
                value:
                  booking_id: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Email sending failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    SlotDTO:
      type: object
      description: A time slot (available, booked, or blocked)
      properties:
        id:
          type: string
          format: uuid
          description: Unique slot identifier
        studio_id:
          type: string
          description: Studio identifier (rishon or ashdod)
        start_at:
          type: string
          format: date-time
          description: Slot start time (ISO 8601 with timezone)
        end_at:
          type: string
          format: date-time
          description: Slot end time (ISO 8601 with timezone)
        status:
          type: string
          enum:
            - available
            - booked
            - blocked
          description: Slot status
      required:
        - id
        - studio_id
        - start_at
        - end_at
        - status
...
    GetSlotsResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/SlotDTO'
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - INVALID_PARAMS
                - SLOT_NOT_FOUND
                - SLOT_UNAVAILABLE
                - SLOT_HAS_ACTIVE_BOOKING
                - BOOKING_NOT_FOUND
                - ALREADY_CANCELLED
                - UNAUTHORIZED
                - VALIDATION_ERROR
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
...
  securitySchemes:
    AdminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin secret key (value of ADMIN_SECRET_KEY env var)

    InternalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Internal secret key (value of ADMIN_SECRET_KEY env var)
...